backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
mysqlclient==2.2.0
pymysql==1.1.0 